+++ src/clojure2d/extra/overlays.clj
@@ -12,16 +12,11 @@
 ;; See example 12 for usage
 
 (ns clojure2d.extra.overlays
-  (:require [clojure2d.core :as core]
+  (:require [clojure2d.core :refer :all]
             [clojure2d.pixels :as p]
             [clojure2d.color :as c]
             [clojure2d.math :as m]
-            [clojure2d.math.random :as r]
-            [clojure2d.math.vector :as v])
-  (:import [java.awt.image BufferedImage]
-           [java.awt Color]
-           [clojure2d.pixels Pixels]
-           [clojure2d.math.vector Vec2]))
+            [clojure2d.math.random :as r]))
 
 (set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
@@ -32,7 +27,7 @@
 
 (defn- blend-shift-and-add-f
   "Slightly shift channels"
-  [ch ^Pixels p1 ^Pixels p2 x y]
+  [ch p1 p2 x y]
   (let [c1 (p/get-value p1 ch x y)
         c2 (p/get-value p2 ch (dec ^long x) y)]
     (c/blend-values add-compose c1 c2)))
@@ -44,12 +39,12 @@
    (for [y (range 0 h 3)]
      (let [y+ (inc ^long y)
            y++ (inc y+)]
-       (core/set-color canvas (r/irand 180 200) 0 0 42)
-       (core/line canvas 0 y w y)
-       (core/set-color canvas 0 (r/irand 180 200) 0 42)
-       (core/line canvas 0 y+ w y+)
-       (core/set-color canvas 0 0 (r/irand 180 200) 42)
-       (core/line canvas 0 y++ w y++))))
+       (set-color canvas (r/irand 180 200) 0 0 42)
+       (line canvas 0 y w y)
+       (set-color canvas 0 (r/irand 180 200) 0 42)
+       (line canvas 0 y+ w y+)
+       (set-color canvas 0 0 (r/irand 180 200) 42)
+       (line canvas 0 y++ w y++))))
   canvas)
 
 (def tinter1 (partial p/filter-channels (p/make-tint-filter 255 142 25)))
@@ -57,22 +52,22 @@
 
 (defn render-rgb-scanlines
   "Blurs and renders rgb stripes on the image, returns new image. Scale parameter (default 1.6) controls amount of blur. Resulting image is sligtly lighter and desaturated. Correct with normalize filter if necessary."
-  ([^BufferedImage p ^double scale]
-   (let [^long w (core/width p)
-         ^long h (core/height p)
+  ([p {:keys [^double scale] :or {scale 1.6}}]
+   (let [^long w (width p)
+         ^long h (height p)
          rimg (-> p
-                  (core/resize-image (int (/ w scale)) (int (/ h scale)))
-                  (core/resize-image w h)
+                  (resize-image (int (/ w scale)) (int (/ h scale)))
+                  (resize-image w h)
                   (p/get-image-pixels))
          l1 (tinter1 rimg)
          l2 (tinter2 rimg)
-         canvas (core/with-canvas (core/create-canvas w h)
-                  (core/image (p/image-from-pixels l1))
+         canvas (with-canvas (create-canvas w h)
+                  (image (p/image-from-pixels l1))
                   (draw-lines w h))]
      
      (let [l1 (p/get-canvas-pixels canvas)]
        (p/image-from-pixels (p/blend-channels (partial p/blend-channel-xy blend-shift-and-add-f) l1 l2)))))
-  ([p] (render-rgb-scanlines p 1.6)))
+  ([p] (render-rgb-scanlines p {})))
 
 ;; ## CRT Scanlines
 ;;
@@ -97,10 +92,10 @@
   * mask-light - crt mask color part multiplier 1.0-1.75 (default 1.0, none)
   * mask-mult - crt mask pattern shift, 0.0+ (default 3.0, crt grid)"
   ([img] (render-crt-scanlines img {}))
-  ([^BufferedImage img {:keys [^double resolution ^double hardpix ^double hardscan ^double mask-dark ^double mask-light ^int mask-mult]
-                        :or {resolution 6.0 hardpix -4.0 hardscan -12.0 mask-dark 1.0 mask-light 1.0 mask-mult 3.0}}]
-   (let [^int w (core/width img)
-         ^int h (core/height img)
+  ([img {:keys [^double resolution ^double hardpix ^double hardscan ^double mask-dark ^double mask-light ^int mask-mult]
+         :or {resolution 6.0 hardpix -4.0 hardscan -12.0 mask-dark 1.0 mask-light 1.0 mask-mult 3.0}}]
+   (let [^int w (width img)
+         ^int h (height img)
          p (p/get-image-pixels img)]
 
      (letfn [(dist [^double pos]
@@ -155,26 +150,27 @@
 
 (defn make-noise
   "Create noise image with set alpha channel (first parameter)."
-  [alpha w h]
-  (let [fc (fn [v] 
-             (c/clamp255 (+ 100.0 (* 20.0 (r/grand)))))
-        fa (fn [v] alpha)
-        p (p/filter-channels (partial p/filter-channel fc) nil nil (partial p/filter-channel fa) (p/make-pixels w h))]
-    (p/set-channel p 1 (p/get-channel p 0))
-    (p/set-channel p 2 (p/get-channel p 0))
-    (p/image-from-pixels p)))
+  ([w h {:keys [alpha] :or {alpha 80}}]
+   (let [fc (fn [v] 
+              (c/clamp255 (+ 100.0 (* 20.0 (r/grand)))))
+         fa (fn [v] alpha)
+         p (p/filter-channels (partial p/filter-channel fc) nil nil (partial p/filter-channel fa) (p/make-pixels w h))]
+     (p/set-channel p 1 (p/get-channel p 0))
+     (p/set-channel p 2 (p/get-channel p 0))
+     (p/image-from-pixels p)))
+  ([w h] (make-noise w h {})))
 
 (defn render-noise
   "Render noise on image"
-  ([noise ^BufferedImage img]
-   (let [w (core/width img)
-         h (core/height img)
-         canvas (core/with-canvas (core/create-canvas w h)
-                  (core/image img)
-                  (core/image noise))]
-     (core/get-image canvas)))
-  ([^BufferedImage img]
-   (render-noise (make-noise 80 (core/width img) (core/height img)) img)))
+  ([img noise]
+   (let [w (width img)
+         h (height img)
+         canvas (with-canvas (create-canvas w h)
+                  (image img)
+                  (image noise))]
+     (get-image canvas)))
+  ([img]
+   (render-noise img (make-noise (width img) (height img)))))
 
 ;; ## Spots
 ;;
@@ -206,7 +202,7 @@
                                                (int))]
                                      (aset pc (+ ^long m (* w ^long n)) bc)
                                      (aset pa (+ ^long m (* w ^long n)) a)))))))
-    (let [^Pixels p (p/make-pixels w h)]
+    (let [p (p/make-pixels w h)]
       (p/set-channel p 0 pc)
       (p/set-channel p 3 pa)
       (let [res (p/filter-channels p/dilate-filter nil nil p/dilate-filter p)]
@@ -216,27 +212,25 @@
 
 (defn make-spots
   "Create vector of spotted overlays. Input: spots transparency (default 80), list of intensities (int values from 0 to 255, default [60 120]) and size of overlay."
-  ([alpha intensities w h]
+  ([w h {:keys [alpha intensities] :or {alpha 80 intensities [60 120]}}]
    (mapv #(spots alpha % w h) intensities))
-  ([w h] (make-spots 80 [60 120] w h)))
+  ([w h] (make-spots w h {})))
 
 (defn- apply-images
   "Add all spotted overlays to image."
   [canvas img spots]
-  (core/image canvas img)
-  (doseq [^BufferedImage s spots]
-    (core/image canvas s))
+  (image canvas img)
+  (doseq [s spots]
+    (image canvas s))
   canvas)
 
 (defn render-spots
   "Render spots on image. Returns image."
-  ([alpha intensities ^BufferedImage img]
-   (let [spots (make-spots alpha intensities (core/width img) (core/height img))]
-     (render-spots spots img)))
-  ([spots ^BufferedImage img]
-   (let [w (core/width img)
-         h (core/height img)
-         canvas (core/with-canvas (core/create-canvas w h)
+  ([img spots]
+   (let [w (width img)
+         h (height img)
+         canvas (with-canvas (create-canvas w h)
                   (apply-images img spots))]     
-     (core/get-image canvas))))
+     (get-image canvas)))
+  ([img] (render-spots img (make-spots (width img) (height img)))))